add double point position

+++ src/main/java/org/example/Point.java
@@ -1,10 +1,10 @@
 package org.example;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
